refactor: jsonNaming 제거
