+++ hicc-reboot/src/main/java/hiccreboot/backend/dto/request/SignUpRequest.java
@@ -1,8 +1,5 @@
 package hiccreboot.backend.dto.request;
 
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
 import hiccreboot.backend.domain.Department;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
@@ -14,7 +11,6 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @Schema(title = "회원가입 요청 DTO")
 public class SignUpRequest {
 	@NotBlank
