+++ hicc-reboot/src/main/java/hiccreboot/backend/dto/request/StudentNumberCheckRequest.java
@@ -1,8 +1,5 @@
 package hiccreboot.backend.dto.request;
 
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import lombok.AccessLevel;
@@ -11,7 +8,6 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class StudentNumberCheckRequest {
 
 	@NotBlank
