+++ hicc-reboot/src/main/java/hiccreboot/backend/dto/response/MemberSimpleResponse.java
@@ -1,8 +1,5 @@
 package hiccreboot.backend.dto.response;
 
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
 import hiccreboot.backend.domain.Grade;
 import hiccreboot.backend.domain.Member;
 import lombok.AccessLevel;
@@ -10,7 +7,6 @@ import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class MemberSimpleResponse {
 	private final Long id;
 	private final String name;
